% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_api_url.R
\name{base_api_url}
\alias{base_api_url}
\title{base_api_url}
\usage{
base_api_url(environment = c("prod", "dev"))
}
\arguments{
\item{environment}{= either "prod" or "dev"; default = "prod".}
}
\value{
a [`character`] string.
}
\description{
Create the URL of the API for either the "prod" or "dev"
environment in the World Inequality Database.
}
\examples{
bau1 <- base_api_url()
bau0 <- base_api_url('dev')
}
\author{
Spencer Graves (replacing a variable by Thomas Blanchet)
}
