% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-args.R
\name{check_ages}
\alias{check_ages}
\title{Check list of age codes}
\usage{
check_ages(ages)
}
\arguments{
\item{ages}{List of age codes}
}
\description{
Check that the list of age codes submitted by the
user is valid.
}
\author{
Thomas Blanchet
}
