% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-args.R
\name{check_areas}
\alias{check_areas}
\title{Check list of area codes}
\usage{
check_areas(areas)
}
\arguments{
\item{areas}{List of area codes}
}
\description{
Check that the list of area codes submitted by the
user is valid.
}
\author{
Thomas Blanchet
}
