% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-args.R
\name{check_indicators}
\alias{check_indicators}
\title{Check list of indicator codes}
\usage{
check_indicators(indicators)
}
\arguments{
\item{indicators}{List of indicators.}
}
\description{
Check that the list of indicator codes submitted by the
user is valid.
}
\author{
Thomas Blanchet
}
