% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-args.R
\name{check_perc}
\alias{check_perc}
\title{Check list of percentiles}
\usage{
check_perc(perc)
}
\arguments{
\item{perc}{List of percentiles}
}
\description{
Check that the list of percentiles submitted by the user is valid
}
\author{
Thomas Blanchet
}
