% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-args.R
\name{check_pop}
\alias{check_pop}
\title{Check list of population codes}
\usage{
check_pop(pop)
}
\arguments{
\item{pop}{List of population codes}
}
\description{
Check that the list of population codes submitted by the
user is valid.
}
\author{
Thomas Blanchet
}
