% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-args.R
\name{check_years}
\alias{check_years}
\title{Check list of years}
\usage{
check_years(years)
}
\arguments{
\item{years}{List of years}
}
\description{
Check that the list of years submitted by the user is valid
}
\author{
Thomas Blanchet
}
