% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-requests.R
\docType{data}
\name{environment}
\alias{environment}
\title{Get variables associated to a list of area codes}
\format{
An object of class \code{character} of length 1.
}
\usage{
environment
}
\arguments{
\item{areas}{List of area codes.}

\item{sixlet}{Six-letter code for which to fetch variables.}
}
\description{
Perform the GET request to the server to retrieve all
variables for a list of area codes.
}
\author{
Thomas Blanchet
}
\keyword{datasets}
