% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-requests.R
\name{get_metadata_variables}
\alias{get_metadata_variables}
\title{Get metadata associated to a list of variables}
\usage{
get_metadata_variables(
  areas,
  variables,
  report_missing = TRUE,
  collected_metadata = NULL
)
}
\arguments{
\item{areas}{List of area codes.}

\item{variables}{List of variables, of the form: \code{"xxxxxx_pXXpYY_999_i"}}

\item{report_missing}{Logical, whether to report missing metadata (default TRUE)}

\item{collected_metadata}{List of metadata already collected (default NULL)}
}
\description{
Perform GET request to the server to retrieve metadata
associated to a list of variables.
}
\author{
Thomas Blanchet
}
