% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-requests.R
\name{get_variables_areas}
\alias{get_variables_areas}
\title{Get variables associated to a list of area codes}
\usage{
get_variables_areas(areas)
}
\arguments{
\item{areas}{List of area codes.}
}
\description{
Perform the GET request to the server to retrieve all
variables for a list of area codes.
}
\author{
Thomas Blanchet
}
