\name{wid_age_codes}
\alias{wid_age_codes}

\title{Age codes in WID.world}

\description{
Age codes are three-digit codes that refer to an age category.
}


\section{List of concepts}{
\tabular{rcl}{
\bold{three-digit age code} \tab      \tab \bold{description} \cr
\code{999} \tab      \tab the population is comprised of individuals of all ages. \cr
\code{991} \tab      \tab the population is comprised of individuals below age 20. \cr
\code{992} \tab      \tab the population is comprised of individuals over age 20. \cr
\code{993} \tab      \tab the population is comprised of individuals between ages 20 and 39. \cr
\code{994} \tab      \tab the population is comprised of individuals between ages 40 and 59. \cr
\code{995} \tab      \tab the population is comprised of individuals over age 60. \cr
\code{996} \tab      \tab the population is comprised of individuals between ages 20 and 65. \cr
\code{997} \tab      \tab the population is comprised of individuals over age 65. \cr
\code{998} \tab      \tab the population is comprised of individuals over age 80. \cr
— \cr
\code{200} \tab      \tab the population is comprised of individuals of age 20. \cr
\code{210} \tab      \tab the population is comprised of individuals of age 21. \cr
\code{220} \tab      \tab the population is comprised of individuals of age 22. \cr
\code{230} \tab      \tab the population is comprised of individuals of age 23. \cr
\code{240} \tab      \tab the population is comprised of individuals of age 24. \cr
\code{250} \tab      \tab the population is comprised of individuals of age 25. \cr
\code{260} \tab      \tab the population is comprised of individuals of age 26. \cr
\code{270} \tab      \tab the population is comprised of individuals of age 27. \cr
\code{280} \tab      \tab the population is comprised of individuals of age 28. \cr
\code{290} \tab      \tab the population is comprised of individuals of age 29. \cr
\code{300} \tab      \tab the population is comprised of individuals of age 30. \cr
\code{310} \tab      \tab the population is comprised of individuals of age 31. \cr
\code{320} \tab      \tab the population is comprised of individuals of age 32. \cr
\code{330} \tab      \tab the population is comprised of individuals of age 33. \cr
\code{340} \tab      \tab the population is comprised of individuals of age 34. \cr
\code{350} \tab      \tab the population is comprised of individuals of age 35. \cr
\code{360} \tab      \tab the population is comprised of individuals of age 36. \cr
\code{370} \tab      \tab the population is comprised of individuals of age 37. \cr
\code{380} \tab      \tab the population is comprised of individuals of age 38. \cr
\code{390} \tab      \tab the population is comprised of individuals of age 39. \cr
\code{400} \tab      \tab the population is comprised of individuals of age 40. \cr
\code{410} \tab      \tab the population is comprised of individuals of age 41. \cr
\code{420} \tab      \tab the population is comprised of individuals of age 42. \cr
\code{430} \tab      \tab the population is comprised of individuals of age 43. \cr
\code{440} \tab      \tab the population is comprised of individuals of age 44. \cr
\code{450} \tab      \tab the population is comprised of individuals of age 45. \cr
\code{460} \tab      \tab the population is comprised of individuals of age 46. \cr
\code{470} \tab      \tab the population is comprised of individuals of age 47. \cr
\code{480} \tab      \tab the population is comprised of individuals of age 48. \cr
\code{490} \tab      \tab the population is comprised of individuals of age 49. \cr
\code{500} \tab      \tab the population is comprised of individuals of age 50. \cr
\code{510} \tab      \tab the population is comprised of individuals of age 51. \cr
\code{520} \tab      \tab the population is comprised of individuals of age 52. \cr
\code{530} \tab      \tab the population is comprised of individuals of age 53. \cr
\code{540} \tab      \tab the population is comprised of individuals of age 54. \cr
\code{550} \tab      \tab the population is comprised of individuals of age 55. \cr
\code{560} \tab      \tab the population is comprised of individuals of age 56. \cr
\code{570} \tab      \tab the population is comprised of individuals of age 57. \cr
\code{580} \tab      \tab the population is comprised of individuals of age 58. \cr
\code{590} \tab      \tab the population is comprised of individuals of age 59. \cr
\code{600} \tab      \tab the population is comprised of individuals of age 60. \cr
\code{610} \tab      \tab the population is comprised of individuals of age 61. \cr
\code{620} \tab      \tab the population is comprised of individuals of age 62. \cr
\code{630} \tab      \tab the population is comprised of individuals of age 63. \cr
\code{640} \tab      \tab the population is comprised of individuals of age 64. \cr
\code{650} \tab      \tab the population is comprised of individuals of age 65. \cr
\code{660} \tab      \tab the population is comprised of individuals of age 66. \cr
\code{670} \tab      \tab the population is comprised of individuals of age 67. \cr
\code{680} \tab      \tab the population is comprised of individuals of age 68. \cr
\code{690} \tab      \tab the population is comprised of individuals of age 69. \cr
\code{700} \tab      \tab the population is comprised of individuals of age 70. \cr
\code{710} \tab      \tab the population is comprised of individuals of age 71. \cr
\code{720} \tab      \tab the population is comprised of individuals of age 72. \cr
\code{730} \tab      \tab the population is comprised of individuals of age 73. \cr
\code{740} \tab      \tab the population is comprised of individuals of age 74. \cr
\code{750} \tab      \tab the population is comprised of individuals of age 75. \cr
\code{760} \tab      \tab the population is comprised of individuals of age 76. \cr
\code{770} \tab      \tab the population is comprised of individuals of age 77. \cr
\code{780} \tab      \tab the population is comprised of individuals of age 78. \cr
\code{790} \tab      \tab the population is comprised of individuals of age 79. \cr
\code{800} \tab      \tab the population is comprised of individuals of age 80. \cr
\code{810} \tab      \tab the population is comprised of individuals of age 81. \cr
\code{820} \tab      \tab the population is comprised of individuals of age 82. \cr
\code{830} \tab      \tab the population is comprised of individuals of age 83. \cr
\code{840} \tab      \tab the population is comprised of individuals of age 84. \cr
\code{850} \tab      \tab the population is comprised of individuals of age 85. \cr
\code{860} \tab      \tab the population is comprised of individuals of age 86. \cr
\code{870} \tab      \tab the population is comprised of individuals of age 87. \cr
\code{880} \tab      \tab the population is comprised of individuals of age 88. \cr
\code{890} \tab      \tab the population is comprised of individuals of age 89. \cr
\code{900} \tab      \tab the population is comprised of individuals of age 90. \cr
\code{910} \tab      \tab the population is comprised of individuals of age 91. \cr
\code{920} \tab      \tab the population is comprised of individuals of age 92. \cr
\code{930} \tab      \tab the population is comprised of individuals of age 93. \cr
\code{940} \tab      \tab the population is comprised of individuals of age 94. \cr
\code{950} \tab      \tab the population is comprised of individuals of age 95. \cr
\code{960} \tab      \tab the population is comprised of individuals of age 96. \cr
\code{970} \tab      \tab the population is comprised of individuals of age 97. \cr
\code{980} \tab      \tab the population is comprised of individuals of age 98. \cr
\code{990} \tab      \tab the population is comprised of individuals of age 99 and above. \cr
— \cr
\code{201} \tab      \tab the population is comprised of individuals of in the 20 to 24 age group. \cr
\code{251} \tab      \tab the population is comprised of individuals of in the 25 to 29 age group. \cr
\code{301} \tab      \tab the population is comprised of individuals of in the 30 to 34 age group. \cr
\code{351} \tab      \tab the population is comprised of individuals of in the 35 to 39 age group. \cr
\code{401} \tab      \tab the population is comprised of individuals of in the 40 to 44 age group. \cr
\code{451} \tab      \tab the population is comprised of individuals of in the 45 to 49 age group. \cr
\code{501} \tab      \tab the population is comprised of individuals of in the 50 to 54 age group. \cr
\code{551} \tab      \tab the population is comprised of individuals of in the 55 to 59 age group. \cr
\code{601} \tab      \tab the population is comprised of individuals of in the 60 to 64 age group. \cr
\code{651} \tab      \tab the population is comprised of individuals of in the 65 to 69 age group. \cr
\code{701} \tab      \tab the population is comprised of individuals of in the 70 to 74 age group. \cr
\code{751} \tab      \tab the population is comprised of individuals of in the 75 to 79 age group. \cr
\code{801} \tab      \tab the population is comprised of individuals of in the 80 to 84 age group. \cr
\code{851} \tab      \tab the population is comprised of individuals of in the 85 to 89 age group. \cr
\code{901} \tab      \tab the population is comprised of individuals of in the 90 to 94 age group. \cr
\code{951} \tab      \tab the population is comprised of individuals of in the 95 to 99 age group. \cr
\code{111} \tab      \tab the population is comprised of individuals over age 99. \cr
— \cr
\code{001} \tab      \tab the population is comprised of individuals of in the 0 to 4 age group. \cr
\code{051} \tab      \tab the population is comprised of individuals of in the 5 to 9 age group. \cr
\code{101} \tab      \tab the population is comprised of individuals of in the 10 to 14 age group. \cr
\code{151} \tab      \tab the population is comprised of individuals of in the 15 to 19 age group. \cr
\code{202} \tab      \tab the population is comprised of individuals of in the 20 to 29 age group. \cr
\code{302} \tab      \tab the population is comprised of individuals of in the 30 to 39 age group. \cr
\code{402} \tab      \tab the population is comprised of individuals of in the 40 to 49 age group. \cr
\code{502} \tab      \tab the population is comprised of individuals of in the 50 to 59 age group. \cr
\code{602} \tab      \tab the population is comprised of individuals of in the 60 to 69 age group. \cr
\code{702} \tab      \tab the population is comprised of individuals of in the 70 to 79 age group. \cr
\code{802} \tab      \tab the population is comprised of individuals of in the 80 to 89 age group. \cr
\code{902} \tab      \tab the population is comprised of individuals of in the 90 to 99 age group. \cr

}
}

\seealso{
\code{\link{download_wid}}
}

\author{
Thomas Blanchet
}
