\name{wid_area_codes}
\alias{wid_area_codes}

\title{Area codes in WID.world}

\description{
Area codes correspond to a country, region within a country, a reunion of countries (e.g. a continent), or even the whole world. For countries and macro-regions, they are two-letter codes, and for countries they the two-letter ISO codes. Regions within a country are note \code{XX-YY} where \code{XX} is the country code and \code{YY} is the region code.
}


\section{List of concepts}{

\subsection{Countries}{
\tabular{rcl}{
\bold{two-letter country code} \tab      \tab \bold{description} \cr
\code{AD} \tab      \tab Andorra \cr
\code{AE} \tab      \tab United Arab Emirates \cr
\code{AF} \tab      \tab Afghanistan \cr
\code{AG} \tab      \tab Antigua and Barbuda \cr
\code{AI} \tab      \tab Anguilla \cr
\code{AL} \tab      \tab Albania \cr
\code{AM} \tab      \tab Armenia \cr
\code{AN} \tab      \tab Netherlands Antilles \cr
\code{AO} \tab      \tab Angola \cr
\code{AR} \tab      \tab Argentina \cr
\code{AS} \tab      \tab American Samoa \cr
\code{AT} \tab      \tab Austria \cr
\code{AU} \tab      \tab Australia \cr
\code{AW} \tab      \tab Aruba \cr
\code{AX} \tab      \tab Aland Islands \cr
\code{AZ} \tab      \tab Azerbaijan \cr
\code{BA} \tab      \tab Bosnia and Herzegovina \cr
\code{BB} \tab      \tab Barbados \cr
\code{BD} \tab      \tab Bangladesh \cr
\code{BE} \tab      \tab Belgium \cr
\code{BF} \tab      \tab Burkina Faso \cr
\code{BG} \tab      \tab Bulgaria \cr
\code{BH} \tab      \tab Bahrain \cr
\code{BI} \tab      \tab Burundi \cr
\code{BJ} \tab      \tab Benin \cr
\code{BL} \tab      \tab Saint Barthelemy \cr
\code{BM} \tab      \tab Bermuda \cr
\code{BN} \tab      \tab Brunei Darussalam \cr
\code{BO} \tab      \tab Bolivia \cr
\code{BQ} \tab      \tab Bonaire, Sint Eustatius and Saba \cr
\code{BR} \tab      \tab Brazil \cr
\code{BS} \tab      \tab Bahamas \cr
\code{BT} \tab      \tab Bhutan \cr
\code{BW} \tab      \tab Botswana \cr
\code{BY} \tab      \tab Belarus \cr
\code{BZ} \tab      \tab Belize \cr
\code{CA} \tab      \tab Canada \cr
\code{CD} \tab      \tab DR Congo \cr
\code{CF} \tab      \tab Central African Republic \cr
\code{CG} \tab      \tab Congo \cr
\code{CH} \tab      \tab Switzerland \cr
\code{CI} \tab      \tab Cote d'Ivoire \cr
\code{CK} \tab      \tab Cook Islands \cr
\code{CL} \tab      \tab Chile \cr
\code{CM} \tab      \tab Cameroon \cr
\code{CN} \tab      \tab China \cr
\code{CO} \tab      \tab Colombia \cr
\code{CR} \tab      \tab Costa Rica \cr
\code{CS} \tab      \tab Czechoslovakia \cr
\code{CU} \tab      \tab Cuba \cr
\code{CV} \tab      \tab Cabo Verde \cr
\code{CW} \tab      \tab Curacao \cr
\code{CY} \tab      \tab Cyprus \cr
\code{CZ} \tab      \tab Czech Republic \cr
\code{DD} \tab      \tab German Democratic Republic \cr
\code{DE} \tab      \tab Germany \cr
\code{DJ} \tab      \tab Djibouti \cr
\code{DK} \tab      \tab Denmark \cr
\code{DM} \tab      \tab Dominica \cr
\code{DO} \tab      \tab Dominican Republic \cr
\code{DZ} \tab      \tab Algeria \cr
\code{EC} \tab      \tab Ecuador \cr
\code{EE} \tab      \tab Estonia \cr
\code{EG} \tab      \tab Egypt \cr
\code{EH} \tab      \tab Western Sahara \cr
\code{ER} \tab      \tab Eritrea \cr
\code{ES} \tab      \tab Spain \cr
\code{ET} \tab      \tab Ethiopia \cr
\code{FI} \tab      \tab Finland \cr
\code{FJ} \tab      \tab Fiji \cr
\code{FK} \tab      \tab Falkland Islands \cr
\code{FM} \tab      \tab Micronesia \cr
\code{FO} \tab      \tab Faroe Islands \cr
\code{FR} \tab      \tab France \cr
\code{GA} \tab      \tab Gabon \cr
\code{GB} \tab      \tab United Kingdom \cr
\code{GD} \tab      \tab Grenada \cr
\code{GE} \tab      \tab Georgia \cr
\code{GG} \tab      \tab Guernsey \cr
\code{GH} \tab      \tab Ghana \cr
\code{GI} \tab      \tab Gibraltar \cr
\code{GL} \tab      \tab Greenland \cr
\code{GM} \tab      \tab Gambia \cr
\code{GN} \tab      \tab Guinea \cr
\code{GP} \tab      \tab Guadeloupe \cr
\code{GQ} \tab      \tab Equatorial Guinea \cr
\code{GR} \tab      \tab Greece \cr
\code{GT} \tab      \tab Guatemala \cr
\code{GU} \tab      \tab Guam \cr
\code{GW} \tab      \tab Guinea-Bissau \cr
\code{GY} \tab      \tab Guyana \cr
\code{HK} \tab      \tab Hong Kong \cr
\code{HN} \tab      \tab Honduras \cr
\code{HR} \tab      \tab Croatia \cr
\code{HT} \tab      \tab Haiti \cr
\code{HU} \tab      \tab Hungary \cr
\code{ID} \tab      \tab Indonesia \cr
\code{IE} \tab      \tab Ireland \cr
\code{IL} \tab      \tab Israel \cr
\code{IM} \tab      \tab Isle of Man \cr
\code{IN} \tab      \tab India \cr
\code{IQ} \tab      \tab Iraq \cr
\code{IR} \tab      \tab Iran \cr
\code{IS} \tab      \tab Iceland \cr
\code{IT} \tab      \tab Italy \cr
\code{JE} \tab      \tab Jersey \cr
\code{JM} \tab      \tab Jamaica \cr
\code{JO} \tab      \tab Jordan \cr
\code{JP} \tab      \tab Japan \cr
\code{KE} \tab      \tab Kenya \cr
\code{KG} \tab      \tab Kyrgyzstan \cr
\code{KH} \tab      \tab Cambodia \cr
\code{KI} \tab      \tab Kiribati \cr
\code{KM} \tab      \tab Comoros \cr
\code{KN} \tab      \tab Saint Kitts and Nevis \cr
\code{KP} \tab      \tab North Korea \cr
\code{KR} \tab      \tab Korea \cr
\code{KS} \tab      \tab Kosovo \cr
\code{KW} \tab      \tab Kuwait \cr
\code{KY} \tab      \tab Cayman Islands \cr
\code{KZ} \tab      \tab Kazakhstan \cr
\code{LA} \tab      \tab Lao PDR \cr
\code{LB} \tab      \tab Lebanon \cr
\code{LC} \tab      \tab Saint Lucia \cr
\code{LI} \tab      \tab Liechtenstein \cr
\code{LK} \tab      \tab Sri Lanka \cr
\code{LR} \tab      \tab Liberia \cr
\code{LS} \tab      \tab Lesotho \cr
\code{LT} \tab      \tab Lithuania \cr
\code{LU} \tab      \tab Luxembourg \cr
\code{LV} \tab      \tab Latvia \cr
\code{LY} \tab      \tab Libya \cr
\code{MA} \tab      \tab Morocco \cr
\code{MC} \tab      \tab Monaco \cr
\code{MD} \tab      \tab Moldova \cr
\code{ME} \tab      \tab Montenegro \cr
\code{MF} \tab      \tab Saint Martin (French part) \cr
\code{MG} \tab      \tab Madagascar \cr
\code{MH} \tab      \tab Marshall Islands \cr
\code{MK} \tab      \tab Macedonia \cr
\code{ML} \tab      \tab Mali \cr
\code{MM} \tab      \tab Myanmar \cr
\code{MN} \tab      \tab Mongolia \cr
\code{MO} \tab      \tab Macao \cr
\code{MP} \tab      \tab Northern Mariana Islands \cr
\code{MQ} \tab      \tab Martinique \cr
\code{MR} \tab      \tab Mauritania \cr
\code{MS} \tab      \tab Montserrat \cr
\code{MT} \tab      \tab Malta \cr
\code{MU} \tab      \tab Mauritius \cr
\code{MV} \tab      \tab Maldives \cr
\code{MW} \tab      \tab Malawi \cr
\code{MX} \tab      \tab Mexico \cr
\code{MY} \tab      \tab Malaysia \cr
\code{MZ} \tab      \tab Mozambique \cr
\code{NA} \tab      \tab Namibia \cr
\code{NC} \tab      \tab New Caledonia \cr
\code{NE} \tab      \tab Niger \cr
\code{NF} \tab      \tab Norfolk Island \cr
\code{NG} \tab      \tab Nigeria \cr
\code{NI} \tab      \tab Nicaragua \cr
\code{NL} \tab      \tab Netherlands \cr
\code{NO} \tab      \tab Norway \cr
\code{NP} \tab      \tab Nepal \cr
\code{NR} \tab      \tab Nauru \cr
\code{NU} \tab      \tab Niue \cr
\code{NZ} \tab      \tab New Zealand \cr
\code{OM} \tab      \tab Oman \cr
\code{PA} \tab      \tab Panama \cr
\code{PE} \tab      \tab Peru \cr
\code{PF} \tab      \tab French Polynesia \cr
\code{PG} \tab      \tab Papua New Guinea \cr
\code{PH} \tab      \tab Philippines \cr
\code{PK} \tab      \tab Pakistan \cr
\code{PL} \tab      \tab Poland \cr
\code{PM} \tab      \tab Saint Pierre and Miquelon \cr
\code{PN} \tab      \tab Pitcairn \cr
\code{PR} \tab      \tab Puerto Rico \cr
\code{PS} \tab      \tab Palestine \cr
\code{PT} \tab      \tab Portugal \cr
\code{PW} \tab      \tab Palau \cr
\code{PY} \tab      \tab Paraguay \cr
\code{QA} \tab      \tab Qatar \cr
\code{RE} \tab      \tab Reunion \cr
\code{RO} \tab      \tab Romania \cr
\code{RS} \tab      \tab Serbia \cr
\code{RU} \tab      \tab Russian Federation \cr
\code{RW} \tab      \tab Rwanda \cr
\code{SA} \tab      \tab Saudi Arabia \cr
\code{SB} \tab      \tab Solomon Islands \cr
\code{SC} \tab      \tab Seychelles \cr
\code{SD} \tab      \tab Sudan \cr
\code{SE} \tab      \tab Sweden \cr
\code{SG} \tab      \tab Singapore \cr
\code{SH} \tab      \tab Saint Helena \cr
\code{SI} \tab      \tab Slovenia \cr
\code{SJ} \tab      \tab Svalbard and Jan Mayen \cr
\code{SK} \tab      \tab Slovakia \cr
\code{SL} \tab      \tab Sierra Leone \cr
\code{SM} \tab      \tab San Marino \cr
\code{SN} \tab      \tab Senegal \cr
\code{SO} \tab      \tab Somalia \cr
\code{SR} \tab      \tab Suriname \cr
\code{SS} \tab      \tab South Sudan \cr
\code{ST} \tab      \tab Sao Tome and Principe \cr
\code{SU} \tab      \tab USSR \cr
\code{SV} \tab      \tab El Salvador \cr
\code{SX} \tab      \tab Sint Maarten (Dutch part) \cr
\code{SY} \tab      \tab Syrian Arab Republic \cr
\code{SZ} \tab      \tab Swaziland \cr
\code{TC} \tab      \tab Turks and Caicos Islands \cr
\code{TD} \tab      \tab Chad \cr
\code{TG} \tab      \tab Togo \cr
\code{TH} \tab      \tab Thailand \cr
\code{TJ} \tab      \tab Tajikistan \cr
\code{TK} \tab      \tab Tokelau \cr
\code{TL} \tab      \tab Timor-Leste \cr
\code{TM} \tab      \tab Turkmenistan \cr
\code{TN} \tab      \tab Tunisia \cr
\code{TO} \tab      \tab Tonga \cr
\code{TR} \tab      \tab Turkey \cr
\code{TT} \tab      \tab Trinidad and Tobago \cr
\code{TV} \tab      \tab Tuvalu \cr
\code{TW} \tab      \tab Taiwan \cr
\code{TZ} \tab      \tab Tanzania \cr
\code{UA} \tab      \tab Ukraine \cr
\code{UG} \tab      \tab Uganda \cr
\code{US} \tab      \tab USA \cr
\code{UY} \tab      \tab Uruguay \cr
\code{UZ} \tab      \tab Uzbekistan \cr
\code{VA} \tab      \tab Holy See \cr
\code{VC} \tab      \tab Saint Vincent and the Grenadines \cr
\code{VE} \tab      \tab Venezuela \cr
\code{VG} \tab      \tab Virgin Islands, British \cr
\code{VI} \tab      \tab Virgin Islands, US \cr
\code{VN} \tab      \tab Viet Nam \cr
\code{VU} \tab      \tab Vanuatu \cr
\code{WF} \tab      \tab Wallis and Futuna \cr
\code{WO} \tab      \tab World \cr
\code{WS} \tab      \tab Samoa \cr
\code{XI} \tab      \tab Channel Islands \cr
\code{YA} \tab      \tab Yemen Arab Republic \cr
\code{YD} \tab      \tab Yemen People's Republic \cr
\code{YE} \tab      \tab Yemen \cr
\code{YT} \tab      \tab Mayotte \cr
\code{YU} \tab      \tab Yugoslavia \cr
\code{ZA} \tab      \tab South Africa \cr
\code{ZM} \tab      \tab Zambia \cr
\code{ZW} \tab      \tab Zimbabwe \cr
\code{ZZ} \tab      \tab Zanzibar \cr
}
}

\subsection{World regions}{
\tabular{rcl}{
\bold{two-letter region code} \tab      \tab \bold{description} \cr
\code{QB} \tab      \tab Africa \cr
\code{QC} \tab      \tab Americas \cr
\code{QD} \tab      \tab Asia \cr
\code{QE} \tab      \tab Europe \cr
\code{QF} \tab      \tab Oceania \cr
\code{QG} \tab      \tab Australia and New Zealand \cr
\code{QH} \tab      \tab Caribbean \cr
\code{QI} \tab      \tab Central America \cr
\code{QJ} \tab      \tab Central Asia \cr
\code{QK} \tab      \tab Eastern Africa \cr
\code{QL} \tab      \tab Eastern Asia \cr
\code{QM} \tab      \tab Eastern Europe \cr
\code{QN} \tab      \tab Middle Africa \cr
\code{QO} \tab      \tab Northern Africa \cr
\code{QP} \tab      \tab Northern America \cr
\code{QQ} \tab      \tab Oceania (excl. Australia and New Zealand) \cr
\code{QR} \tab      \tab South America \cr
\code{QS} \tab      \tab South-Eastern Asia \cr
\code{QT} \tab      \tab Southern Africa \cr
\code{QU} \tab      \tab Southern Asia \cr
\code{QV} \tab      \tab Western Africa \cr
\code{QW} \tab      \tab Western Asia \cr
\code{QX} \tab      \tab Western Europe \cr
\code{QY} \tab      \tab European Union \cr
}
}

\subsection{Country subregions}{
\bold{United States}
\tabular{rcl}{
\bold{five-letter region code} \tab      \tab \bold{description} \cr
\code{US-AL} \tab      \tab Alabama \cr
\code{US-AK} \tab      \tab Alaska \cr
\code{US-AZ} \tab      \tab Arizona \cr
\code{US-AR} \tab      \tab Arkansas \cr
\code{US-CA} \tab      \tab California \cr
\code{US-CO} \tab      \tab Colorado \cr
\code{US-CT} \tab      \tab Connecticut \cr
\code{US-DE} \tab      \tab Delaware \cr
\code{US-DC} \tab      \tab District of Columbia \cr
\code{US-FL} \tab      \tab Florida \cr
\code{US-GA} \tab      \tab Georgia \cr
\code{US-HI} \tab      \tab Hawaii \cr
\code{US-ID} \tab      \tab Idaho \cr
\code{US-IL} \tab      \tab Illinois \cr
\code{US-IN} \tab      \tab Indiana \cr
\code{US-IA} \tab      \tab Iowa \cr
\code{US-KS} \tab      \tab Kansas \cr
\code{US-KY} \tab      \tab Kentucky \cr
\code{US-LA} \tab      \tab Louisiana \cr
\code{US-ME} \tab      \tab Maine \cr
\code{US-MD} \tab      \tab Maryland \cr
\code{US-MA} \tab      \tab Massachusetts \cr
\code{US-MI} \tab      \tab Michigan \cr
\code{US-MN} \tab      \tab Minnesota \cr
\code{US-MS} \tab      \tab Mississippi \cr
\code{US-MO} \tab      \tab Missouri \cr
\code{US-MT} \tab      \tab Montana \cr
\code{US-NE} \tab      \tab Nebraska \cr
\code{US-NV} \tab      \tab Nevada \cr
\code{US-NH} \tab      \tab New Hampshire \cr
\code{US-NJ} \tab      \tab New Jersey \cr
\code{US-NM} \tab      \tab New Mexico \cr
\code{US-NY} \tab      \tab New York \cr
\code{US-NC} \tab      \tab North Carolina \cr
\code{US-ND} \tab      \tab North Dakota \cr
\code{US-OH} \tab      \tab Ohio \cr
\code{US-OK} \tab      \tab Oklahoma \cr
\code{US-OR} \tab      \tab Oregon \cr
\code{US-PA} \tab      \tab Pennsylvania \cr
\code{US-RI} \tab      \tab Rhode Island \cr
\code{US-SC} \tab      \tab South Carolina \cr
\code{US-SD} \tab      \tab South Dakota \cr
\code{US-TN} \tab      \tab Tennessee \cr
\code{US-TX} \tab      \tab Texas \cr
\code{US-UT} \tab      \tab Utah \cr
\code{US-VT} \tab      \tab Vermont \cr
\code{US-VA} \tab      \tab Virginia \cr
\code{US-WA} \tab      \tab Washington \cr
\code{US-WV} \tab      \tab West Virginia \cr
\code{US-WI} \tab      \tab Wisconsin \cr
\code{US-WY} \tab      \tab Wyoming \cr
}

\bold{China}
\tabular{rcl}{
\bold{five-letter region code} \tab      \tab \bold{description} \cr
\code{CN-RU} \tab      \tab Rural China \cr
\code{CN-UR} \tab      \tab Urban China \cr
}

}

}

\seealso{
\code{\link{download_wid}}
}

\author{
Thomas Blanchet
}
