\name{wid_population_codes}
\alias{wid_population_codes}

\title{Population codes in WID.world}

\description{
Population codes are one-letter codes that refer to a definition of the population of interest (i.e. the statistical unit).
}


\section{List of concepts}{
\tabular{rcl}{
\bold{one-letter code} \tab      \tab \bold{description} \cr
\code{i} \tab      \tab individuals \cr
\code{j} \tab      \tab equal-split adults (income or wealth divided equally among spouses) \cr
\code{m} \tab      \tab male \cr
\code{f} \tab      \tab female \cr
\code{t} \tab      \tab tax unit \cr
\code{e} \tab      \tab employed \cr
}
}

\seealso{
\code{\link{download_wid}}
}

\author{
Thomas Blanchet
}

