% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-wid.R
\name{download_wid}
\alias{download_wid}
\title{Download data from WID.world}
\usage{
download_wid(indicators = "all", areas = "all", years = "all",
  perc = "all", ages = "all", pop = "all", metadata = FALSE,
  verbose = FALSE)
}
\arguments{
\item{indicators}{List of six-letter strings, or \code{"all"}:
code names of the indicators in the database. Default is \code{"all"} for all
indicators. See 'Details' for more.}

\item{areas}{List of strings, or \code{"all"}: area code names of the
database. \code{"XX"} for countries/regions, \code{"XX-YY"} for subregions.
Default is \code{"all"} for all areas. See 'Details' for more.}

\item{years}{Numerical vector, or \code{"all"}: years to retrieve.
Default is \code{"all"} for all years.}

\item{perc}{List of strings, or \code{"all"}: percentiles take the form
\code{"pXX"} or \code{"pXXpYY"}. Default is \code{"all"} for all percentiles.
See 'Details' for more.}

\item{ages}{Numerical vector, or \code{"all"}: age category codes in the
database. 999 for all ages, 992 for adults. Default is \code{"all"} for all
age categories. See 'Details' for more.}

\item{pop}{List of characters, or \code{"all"}: type of population.
\code{"t"} for tax units, \code{"i"} for individuals. Default is
\code{"all"} for all population types. See 'Details' for more.}

\item{metadata}{Should the function fetch metadata too (ie. variable
descriptions, sources, methodological notes, etc.)? Default is \code{FALSE}.}

\item{verbose}{Should the function indicate the progress of the request?
Default is \code{FALSE}.}
}
\description{
Downloads data from the World Wealth and Income Database
(\url{http://WID.world}) into a \code{data.frame}.
}
\author{
Thomas Blanchet
}
