% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-requests.R
\docType{data}
\name{environment}
\alias{environment}
\title{Get variables associated to a list of area codes}
\format{
An object of class \code{character} of length 1.
}
\usage{
environment
}
\description{
Package API environment selector.
}
\author{
Thomas Blanchet
}
\keyword{datasets}
