% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_age_codes}
\alias{validate_age_codes}
\title{Validate age codes for WID database}
\usage{
validate_age_codes(ages)
}
\arguments{
\item{ages}{Character or numeric vector of age codes}
}
\value{
Character vector of validated 3-digit age codes
}
\description{
Validate age codes for WID database
}
